requirements.txt
absl-py==2.1.0
accelerate==1.3.0
annotated-types==0.7.0
anyio==4.8.0
astunparse==1.6.3
attrs==25.1.0
audioread==3.0.1
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.4
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
deepface==0.0.93
diskcache==5.6.3
distro==1.9.0
dlib-bin==19.24.6
exceptiongroup==1.2.2
ffmpeg-python==0.2.0
filelock==3.13.1
fire==0.7.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==25.1.24
fonttools==4.56.0
fsspec==2024.6.1
future==1.0.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.70.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
keras==3.8.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llama_cpp_python==0.3.7
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
mdurl==0.1.2
mediapipe==0.10.21
ml-dtypes==0.4.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
mtcnn==0.1.1
namex==0.0.8
networkx==3.2.1
nltk==3.9.1
numba==0.60.0
numpy==1.26.3
openai==1.61.1
openai-whisper==20240930
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
protobuf==4.25.6
psutil==6.1.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.6.1
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
retina-face==0.0.17
rich==13.9.4
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
textblob==0.19.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0

tqdm==4.67.1

typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
